refactor: プロジェクト構造整理とREADME.md更新
- テストファイル(test_*.py)をtests/ディレクトリに移動
- スクリプトファイル(fix_imports.py, start_backend.py)をscripts/に移動
- 生成ファイル(coverage.xml, ruff-backend-check.json)を.gitignoreに追加・削除
- README.mdを大幅更新:
  - Phase3完了(5取引所対応・セキュリティ強化)の成果を反映
  - 技術仕様更新(Python 3.12, Material-UI v5, 98%+カバレッジ)
  - プロジェクト構造図を整理後の実際の構造に更新
  - docs/内の重要ドキュメントへの直接リンク追加
  - 開発完了状況明確化(Phase1-3完了、Phase4計画)
