+++ README.md
@@ -2,42 +2,43 @@
 
 **AI駆動の高度な暗号通貨自動取引システム**
 
-次世代の取引戦略を実装し、リアルタイムでの市場分析と自動取引を実現する包括的なトレーディングプラットフォームです。
+5つの主要取引所対応、包括的なセキュリティシステム、Paper/Live Trading完全分離アーキテクチャを実装した本格的なトレーディングプラットフォームです。
 
 ![License](https://img.shields.io/badge/license-MIT-blue.svg)
-![Python](https://img.shields.io/badge/python-3.11+-blue.svg)
+![Python](https://img.shields.io/badge/python-3.12-blue.svg)
 ![Next.js](https://img.shields.io/badge/next.js-14-black.svg)
 ![FastAPI](https://img.shields.io/badge/fastapi-0.109+-green.svg)
+![Coverage](https://img.shields.io/badge/coverage-98%25-brightgreen.svg)
 
 ## ✨ 主要機能
 
-### 🎯 コア機能
-- **📈 多様な取引戦略**: EMA、RSI、MACD、カスタム戦略対応
-- **💹 リアルタイム取引**: Binance、Bybitと直接連携
-- **🔄 高度なバックテスト**: 歴史的データでの戦略検証
-- **🛡️ リスク管理**: 自動ストップロス、ポジションサイジング
-- **📊 包括的な分析**: パフォーマンス分析とレポート生成
+### 🎯 Phase3 完了機能 (2025年7月)
+- **🏦 5取引所完全対応**: Binance、Bybit、Bitget、Hyperliquid、BackPack Exchange
+- **📊 Paper Trading システム**: 完全な模擬取引環境と資金管理
+- **🛡️ 多層セキュリティ**: JWT+CSRF+レート制限+API キー暗号化
+- **🎨 モダンUI**: Material-UI v5ベースの直感的インターフェース
+- **📈 16種類の取引戦略**: EMA、RSI、MACD、Bollinger Band等の完全実装
 
-### 🆕 Phase 2 新機能 (2025年1月実装)
-- **📡 リアルタイムデータパイプライン**: Binance APIからのOHLCVデータ自動収集
-- **💾 大規模データ管理**: Supabaseへの効率的なバッチ保存（1000件/バッチ）
-- **⚡ 並列処理**: 複数シンボル・タイムフレームの同時データ収集
-- **🔄 自動データ更新**: 定期的な価格データの収集とストレージ
+### 🔧 技術仕様
+- **📡 リアルタイムデータパイプライン**: 並列データ収集とWebSocket価格配信
+- **💾 ハイブリッドDB**: Supabase (PostgreSQL) + DuckDB ローカル高速処理
+- **⚡ 高性能アーキテクチャ**: Factory/Adapter/Repository パターン実装
+- **🧪 包括的テスト**: 98%+ カバレッジ、CI/CD完全自動化
 
 ### 🌐 技術スタック
-- **Backend**: FastAPI (Python 3.11+)
-- **Frontend**: Next.js 14 + React 18
-- **Database**: PostgreSQL (Supabase)
-- **Cache**: Redis
-- **Container**: Docker
-- **CI/CD**: GitHub Actions
+- **Backend**: FastAPI (Python 3.12) + Supabase + DuckDB
+- **Frontend**: Next.js 14 + TypeScript + Material-UI v5
+- **Database**: PostgreSQL (Supabase) + Redis Cache
+- **取引所**: Binance, Bybit, Bitget, Hyperliquid, BackPack
+- **Container**: Docker + Docker Compose
+- **CI/CD**: GitHub Actions (完全自動化)
 
 ## 🚀 5分でスタート
 
 ### 前提条件
-- Node.js 18+
-- Python 3.11+
-- Docker (オプション)
+- Node.js 20+
+- Python 3.12+
+- Docker (推奨)
 
 ### クイックスタート
 
@@ -53,10 +54,10 @@ cp .env.example .env
 # 3. バックエンド起動
 
 ## ローカル開発（フル機能）
-python start_backend.py
+python scripts/start_backend.py
 
 ## または軽量認証テスト
-python test_auth_simple.py  # ポート8001で起動
+python tests/test_auth_simple.py  # ポート8001で起動
 
 # 4. フロントエンド起動（新しいターミナル）
 cd frontend && npm install && npm run dev
@@ -76,31 +77,41 @@ cd frontend && npm install && npm run dev
 |-------------|------|
 | [🚀 Getting Started](./docs/GETTING_STARTED.md) | 5分で始める完全ガイド |
 | [🏗️ Architecture](./docs/ARCHITECTURE.md) | システム設計とアーキテクチャ |
-| [🗺️ Phase2 Roadmap](./docs/PHASE2_ROADMAP.md) | Phase2-3の実装計画 |
+| [🎯 Phase3 完了レポート](./docs/PHASE3_COMPLETION_REPORT.md) | 5取引所対応・セキュリティ強化完了 |
+| [📊 Project Status](./docs/PROJECT_STATUS.md) | 最新開発状況・実装進捗 |
 | [📚 API Reference](./docs/API_REFERENCE.md) | 完全APIリファレンス |
 | [🔧 Contributing Guide](./docs/CONTRIBUTING.md) | 開発参加ガイド |
+| [🛡️ Security Review](./docs/SECURITY_REVIEW_REPORT.md) | セキュリティ監査レポート |
+| [🚀 Future Roadmap](./docs/FUTURE_ROADMAP.md) | Phase4以降の開発計画 |
 
 ## 📊 プロジェクト構造
 
 ```
-crypto-bot/
-├── 📁 backend/              # Python FastAPI サーバー
-│   ├── api/                 # APIエンドポイント
-│   ├── core/                # 共通機能（DB、認証、設定）
-│   ├── strategies/          # 取引戦略実装
-│   ├── exchanges/           # 取引所API連携
-│   ├── backtesting/         # バックテストエンジン
-│   └── data_pipeline/       # データ収集パイプライン
-├── 📁 frontend/             # Next.js フロントエンド
-│   ├── components/          # Reactコンポーネント
-│   ├── pages/              # Next.jsページ
-│   └── store/              # Redux状態管理
-├── 📁 tests/               # テストコード
-├── 📁 docs/                # プロジェクトドキュメント
-└── 📁 .github/             # CI/CDワークフロー
+advanced-crypto-trading-bot/
+├── 📁 src/                   # メインソースコード
+│   ├── backend/              # Python FastAPI サーバー
+│   │   ├── api/             # APIエンドポイント（認証・取引・データ）
+│   │   ├── core/            # 共通機能（DB・認証・セキュリティ）
+│   │   ├── exchanges/       # 5取引所API連携アダプター
+│   │   ├── strategies/      # 16種類の取引戦略実装
+│   │   ├── portfolio/       # ポートフォリオ管理・最適化
+│   │   ├── risk/           # リスク管理・サーキットブレーカー
+│   │   └── streaming/       # WebSocketリアルタイム価格配信
+│   └── vercel_api/          # Vercel Functions専用API
+├── 📁 frontend/             # Next.js 14 + TypeScript + MUI
+│   ├── src/                 # TypeScriptソースコード
+│   │   ├── components/      # Reactコンポーネント（Material-UI）
+│   │   ├── app/            # Next.js App Router ページ
+│   │   ├── store/          # 状態管理（React Context）
+│   │   └── lib/            # ユーティリティ・API客户端
+├── 📁 tests/               # 包括的テストスイート（98%+ カバレッジ）
+├── 📁 scripts/             # 開発・デプロイスクリプト
+├── 📁 docs/                # 完全プロジェクトドキュメント
+├── 📁 config/              # 設定ファイル・戦略パラメータ
+└── 📁 .github/             # CI/CD完全自動化ワークフロー
 ```
 
-## 🎯 現在の開発状況
+## 🎯 開発完了状況
 
 ### ✅ Phase 1: 基盤構築（完了）
 - [x] FastAPI + Next.js アーキテクチャ
@@ -109,19 +120,27 @@ crypto-bot/
 - [x] 基本的な取引戦略実装
 - [x] CI/CDパイプライン構築
 
-### 🚧 Phase 2: データパイプライン（進行中）
-- [x] リアルタイムデータ収集 - Binance OHLCV収集実装完了
-- [x] 取引所API統合強化 - DataCollectorクラス実装
-- [ ] バックテスト機能拡張 - 実データ対応（開発中）
-- [ ] WebSocketリアルタイム通信 - 価格更新機能（計画中）
+### ✅ Phase 2: データパイプライン（完了）
+- [x] リアルタイムデータ収集 - 並列OHLCV収集実装
+- [x] 取引所API統合強化 - 統一Adapterパターン
+- [x] バックテスト機能拡張 - 歴史データ完全対応
+- [x] WebSocketリアルタイム通信 - 価格配信システム
 
-### 📋 Phase 3: 本格運用（計画中）
-- [ ] ライブ取引機能
-- [ ] 高度なリスク管理
+### ✅ Phase 3: 本格運用システム（完了）
+- [x] **5取引所完全対応** - Binance, Bybit, Bitget, Hyperliquid, BackPack
+- [x] **Paper Trading システム** - 完全模擬取引環境
+- [x] **包括的セキュリティ** - 多層防御システム実装
+- [x] **16種類の取引戦略** - 全戦略本格実装・検証済み
+- [x] **Material-UI モダンUI** - 直感的ユーザーインターフェース
+- [x] **98%+ テストカバレッジ** - 品質保証体制完成
+
+### 🚀 Phase 4: 次世代機能（計画中）
 - [ ] ML予測モデル統合
-- [ ] 包括的な監視システム
+- [ ] 高頻度取引（HFT）対応
+- [ ] ソーシャルトレーディング機能
+- [ ] モバイルアプリ開発
 
-> 📈 **詳細な実装内容**: [Phase2 Implementation](./docs/PHASE2_IMPLEMENTATION.md)
+> 📈 **完了レポート**: [Phase3 Completion Report](./docs/PHASE3_COMPLETION_REPORT.md)
 
 ## 🛡️ セキュリティ
