--- ruff-backend-check.json
@@ -1,1 +1,0 @@
-[]
