--- test_backtest_functionality.py
@@ -1,231 +1,0 @@
-#!/usr/bin/env python3
-"""
-バックテスト機能の実動作テスト
-"""
-
-import os
-import sys
-import asyncio
-from pathlib import Path
-from datetime import date
-
-# プロジェクトルートをPythonパスに追加
-project_root = Path(__file__).parent
-sys.path.insert(0, str(project_root))
-
-# ダミー環境変数を設定
-os.environ.setdefault("SUPABASE_URL", "https://dummy.supabase.co")
-os.environ.setdefault("SUPABASE_KEY", "dummy_key")
-os.environ.setdefault("SUPABASE_JWT_SECRET", "dummy_secret")
-os.environ.setdefault("ADMIN_USERNAME", "admin")
-os.environ.setdefault("ALLOWED_ORIGINS", '["http://localhost:3000"]')
-
-
-async def test_backtest_engine_import():
-    """バックテストエンジンのインポートテスト"""
-    try:
-        print("🔍 Testing BacktestEngine import...")
-
-        print("✅ BacktestEngine imported successfully")
-
-        print("✅ Backtest models imported successfully")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Import test failed: {e}")
-        return False
-
-
-async def test_backtest_models():
-    """バックテストモデルの作成テスト"""
-    try:
-        print("\n🔍 Testing BacktestRequest model creation...")
-
-        from backend.api.backtest import BacktestRequest
-
-        # テスト用のリクエストを作成
-        test_request = BacktestRequest(
-            strategy_name="test_strategy",
-            symbol="BTC/USDT",
-            timeframe="1h",
-            start_date=date(2024, 1, 1),
-            end_date=date(2024, 1, 7),
-            initial_capital=10000.0,
-            use_real_data=True,
-            data_quality_threshold=0.95,
-        )
-
-        print("✅ BacktestRequest created successfully")
-        print(f"   - Strategy: {test_request.strategy_name}")
-        print(f"   - Symbol: {test_request.symbol}")
-        print(f"   - Timeframe: {test_request.timeframe}")
-        print(f"   - Date range: {test_request.start_date} to {test_request.end_date}")
-        print(f"   - Initial capital: ${test_request.initial_capital:,}")
-        print(f"   - Use real data: {test_request.use_real_data}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Model test failed: {e}")
-        return False
-
-
-async def test_data_validation_logic():
-    """データバリデーションロジックのテスト"""
-    try:
-        print("\n🔍 Testing data validation logic...")
-
-        from backend.backtesting.engine import DataValidator
-        import pandas as pd
-
-        # テスト用のOHLCVデータを作成
-        test_dates = pd.date_range(start="2024-01-01", periods=24, freq="H")
-        test_data = pd.DataFrame(
-            {
-                "timestamp": test_dates,
-                "open": [50000 + i * 100 for i in range(24)],
-                "high": [50100 + i * 100 for i in range(24)],
-                "low": [49900 + i * 100 for i in range(24)],
-                "close": [50050 + i * 100 for i in range(24)],
-                "volume": [1000 + i * 10 for i in range(24)],
-            }
-        )
-
-        # データ品質レポートを生成
-        report = DataValidator.validate_ohlcv_data(test_data, "BTC/USDT", "1h")
-
-        print("✅ Data validation completed")
-        print(f"   - Total records: {report.total_records}")
-        print(f"   - Missing records: {report.missing_records}")
-        print(f"   - Data coverage: {report.data_coverage:.2%}")
-        print(f"   - Quality score: {report.quality_score:.2f}")
-        print(f"   - Is valid: {report.is_valid()}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Data validation test failed: {e}")
-        return False
-
-
-async def test_strategy_loader():
-    """ストラテジーローダーのテスト"""
-    try:
-        print("\n🔍 Testing strategy loader...")
-
-        from backend.strategies.loader import StrategyLoader
-
-        loader = StrategyLoader()
-
-        # 利用可能なストラテジーを確認
-        available_strategies = loader.get_available_strategies()
-        print("✅ Strategy loader initialized")
-        print(f"   - Available strategies: {len(available_strategies)}")
-
-        for strategy_name in available_strategies[:5]:  # 最初の5つだけ表示
-            print(f"     - {strategy_name}")
-
-        if len(available_strategies) > 5:
-            print(f"     ... and {len(available_strategies) - 5} more")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Strategy loader test failed: {e}")
-        return False
-
-
-async def test_performance_metrics():
-    """パフォーマンス指標計算のテスト"""
-    try:
-        print("\n🔍 Testing performance metrics calculation...")
-
-        import pandas as pd
-        import numpy as np
-
-        # テスト用のリターンデータを作成
-        dates = pd.date_range(start="2024-01-01", periods=100, freq="D")
-        returns = np.random.normal(0.001, 0.02, 100)  # 日次リターン
-
-        test_portfolio = pd.Series(returns, index=dates).cumsum()
-
-        # パフォーマンス指標を計算
-        # シンプルなメトリクス計算のテスト
-        metrics = {
-            "total_return": test_portfolio.iloc[-1] - test_portfolio.iloc[0],
-            "volatility": returns.std(),
-            "sharpe_ratio": returns.mean() / returns.std() if returns.std() > 0 else 0,
-            "max_drawdown": (test_portfolio / test_portfolio.cummax() - 1).min(),
-        }
-
-        print("✅ Performance metrics calculated successfully")
-        print(f"   - Total metrics: {len(metrics)}")
-
-        key_metrics = ["total_return", "sharpe_ratio", "max_drawdown", "win_rate"]
-        for metric in key_metrics:
-            if metric in metrics:
-                value = metrics[metric]
-                if isinstance(value, float):
-                    print(f"   - {metric}: {value:.4f}")
-                else:
-                    print(f"   - {metric}: {value}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Performance metrics test failed: {e}")
-        return False
-
-
-async def main():
-    """メイン実行関数"""
-    print("🚀 Starting comprehensive backtest functionality tests...\n")
-
-    # テスト実行
-    test_results = []
-
-    tests = [
-        ("Backtest Engine Import", test_backtest_engine_import),
-        ("Backtest Models", test_backtest_models),
-        ("Data Validation Logic", test_data_validation_logic),
-        ("Strategy Loader", test_strategy_loader),
-        ("Performance Metrics", test_performance_metrics),
-    ]
-
-    for test_name, test_func in tests:
-        try:
-            result = await test_func()
-            test_results.append((test_name, result))
-        except Exception as e:
-            print(f"❌ Test '{test_name}' encountered an error: {e}")
-            test_results.append((test_name, False))
-
-    # 結果表示
-    print("\n📊 Comprehensive Test Results:")
-    passed = 0
-    for test_name, result in test_results:
-        status = "✅ PASS" if result else "❌ FAIL"
-        print(f"   - {test_name}: {status}")
-        if result:
-            passed += 1
-
-    success_rate = passed / len(test_results) * 100
-    print(f"\n🎯 Success Rate: {passed}/{len(test_results)} ({success_rate:.1f}%)")
-
-    if success_rate >= 80:
-        print("\n🎉 Backtest functionality is working excellently!")
-        print("   Ready for production use with real data integration.")
-    elif success_rate >= 60:
-        print("\n✅ Backtest functionality is mostly working.")
-        print("   Minor issues detected but core functionality is intact.")
-    else:
-        print("\n⚠️  Some critical issues detected.")
-        print("   Please review the failing tests before proceeding.")
-
-    return success_rate >= 80
-
-
-if __name__ == "__main__":
-    success = asyncio.run(main())
-    sys.exit(0 if success else 1)
