--- test_price_streaming.py
@@ -1,312 +1,0 @@
-#!/usr/bin/env python3
-"""
-リアルタイム価格配信システムテスト
-包括的価格ストリーミング機能の動作確認
-"""
-
-import asyncio
-import os
-import sys
-from pathlib import Path
-
-# プロジェクトルートをPythonパスに追加
-project_root = Path(__file__).parent
-sys.path.insert(0, str(project_root))
-
-# ダミー環境変数を設定
-os.environ.setdefault("SUPABASE_URL", "https://dummy.supabase.co")
-os.environ.setdefault("SUPABASE_KEY", "dummy_key")
-os.environ.setdefault("SUPABASE_JWT_SECRET", "dummy_secret")
-os.environ.setdefault("ADMIN_USERNAME", "admin")
-os.environ.setdefault("ALLOWED_ORIGINS", '["http://localhost:3000"]')
-
-
-async def test_price_streamer_import():
-    """価格ストリーマーのインポートテスト"""
-    try:
-        print("🔍 Testing price streamer import...")
-
-        from backend.streaming.price_streamer import PriceData
-
-        print("✅ Price streamer imported successfully")
-
-        # データクラスの確認
-        sample_price = PriceData(
-            symbol="BTCUSDT",
-            price=45000.0,
-            change_24h=1000.0,
-            change_percent_24h=2.27,
-            volume_24h=50000.0,
-            high_24h=46000.0,
-            low_24h=44000.0,
-            timestamp="2024-07-20T10:00:00Z",
-        )
-
-        print("✅ PriceData created successfully")
-        print(f"   - Symbol: {sample_price.symbol}")
-        print(f"   - Price: ${sample_price.price:,}")
-        print(f"   - Change: {sample_price.change_percent_24h:+.2f}%")
-
-        price_dict = sample_price.to_dict()
-        print(f"✅ PriceData serialization: {len(price_dict)} fields")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Import test failed: {e}")
-        return False
-
-
-async def test_price_manager_lifecycle():
-    """価格マネージャーのライフサイクルテスト"""
-    try:
-        print("\n🔍 Testing price manager lifecycle...")
-
-        from backend.streaming.price_streamer import PriceStreamManager
-
-        # テスト用マネージャーを作成
-        test_manager = PriceStreamManager()
-
-        # 初期状態確認
-        initial_stats = test_manager.get_stats()
-        print(f"✅ Initial state: running={initial_stats['manager_running']}")
-
-        # 開始テスト（実際のBinance接続はスキップ）
-        print("✅ Manager lifecycle test completed (mock)")
-        print("   - Start/Stop methods available")
-        print("   - Stats collection working")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Lifecycle test failed: {e}")
-        return False
-
-
-async def test_message_broadcasting():
-    """メッセージブロードキャストテスト"""
-    try:
-        print("\n🔍 Testing message broadcasting...")
-
-        from backend.streaming.price_streamer import PriceData, TradeData
-        from backend.websocket.manager import ChannelType, MessageType, WebSocketMessage
-
-        # 価格データメッセージ
-        sample_price = PriceData(
-            symbol="ETHUSDT",
-            price=3000.0,
-            change_24h=150.0,
-            change_percent_24h=5.26,
-            volume_24h=25000.0,
-            high_24h=3100.0,
-            low_24h=2900.0,
-            timestamp="2024-07-20T10:00:00Z",
-        )
-
-        price_message = WebSocketMessage(
-            type=MessageType.PRICE_UPDATE,
-            channel=ChannelType.PRICES,
-            data=sample_price.to_dict(),
-        )
-
-        print("✅ Price update message created")
-        print(f"   - Type: {price_message.type}")
-        print(f"   - Channel: {price_message.channel}")
-        print(f"   - Symbol: {price_message.data['symbol']}")
-
-        # 取引データメッセージ
-        sample_trade = TradeData(
-            symbol="ADAUSDT",
-            price=0.45,
-            quantity=1000.0,
-            is_buyer_maker=False,
-            timestamp="2024-07-20T10:00:01Z",
-            trade_id="12345",
-        )
-
-        trade_message = WebSocketMessage(
-            type=MessageType.TRADE_EXECUTION,
-            channel=ChannelType.TRADES,
-            data=sample_trade.to_dict(),
-        )
-
-        print("✅ Trade execution message created")
-        print(f"   - Type: {trade_message.type}")
-        print(f"   - Side: {trade_message.data['side']}")
-        print(f"   - Quantity: {trade_message.data['quantity']}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Broadcasting test failed: {e}")
-        return False
-
-
-async def test_streaming_routes_import():
-    """ストリーミングルートのインポートテスト"""
-    try:
-        print("\n🔍 Testing streaming routes import...")
-
-        from backend.streaming.routes import router
-
-        print("✅ Streaming routes imported successfully")
-
-        # ルート確認
-        routes = []
-        for route in router.routes:
-            if hasattr(route, "path"):
-                routes.append(f"{route.methods} {route.path}" if hasattr(route, "methods") else route.path)
-
-        print(f"✅ Found {len(routes)} routes:")
-        for route in routes[:8]:  # 最初の8つだけ表示
-            print(f"   - {route}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Routes import test failed: {e}")
-        return False
-
-
-async def test_main_app_streaming_integration():
-    """main.pyアプリケーションとストリーミング統合テスト"""
-    try:
-        print("\n🔍 Testing main app streaming integration...")
-
-        from backend.main import app
-
-        print("✅ Main app with streaming integration imported")
-
-        # ストリーミングルートが含まれているか確認
-        streaming_routes = []
-        for route in app.routes:
-            if hasattr(route, "path") and "/streaming" in route.path:
-                streaming_routes.append(route.path)
-
-        if streaming_routes:
-            print("✅ Streaming routes found in main app:")
-            for route in streaming_routes:
-                print(f"   - {route}")
-        else:
-            print("⚠️  No streaming routes found in main app")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Main app integration test failed: {e}")
-        return False
-
-
-async def test_binance_streamer_config():
-    """Binanceストリーマー設定テスト"""
-    try:
-        print("\n🔍 Testing Binance streamer configuration...")
-
-        from backend.streaming.price_streamer import BinanceWebSocketStreamer
-
-        # テスト用ストリーマー
-        streamer = BinanceWebSocketStreamer()
-
-        print("✅ Binance streamer initialized")
-        print(f"   - Base URL: {streamer.base_url}")
-        print(f"   - Default symbols: {len(streamer.default_symbols)}")
-        print(f"   - Symbols: {', '.join(streamer.default_symbols[:4])}...")
-
-        # 統計取得テスト
-        stats = streamer.get_connection_stats()
-        print("✅ Connection stats available:")
-        print(f"   - Running: {stats['is_running']}")
-        print(f"   - Subscribed symbols: {stats['subscribed_symbols']}")
-        print(f"   - Active connections: {stats['active_connections']}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ Binance streamer test failed: {e}")
-        return False
-
-
-async def test_websocket_integration():
-    """WebSocket統合テスト"""
-    try:
-        print("\n🔍 Testing WebSocket integration...")
-
-        # WebSocketマネージャーが利用可能か確認
-        from backend.streaming.price_streamer import price_stream_manager
-        from backend.websocket.manager import websocket_manager
-
-        print("✅ WebSocket manager available")
-        print("✅ Price stream manager available")
-
-        # 統計情報の取得
-        ws_stats = websocket_manager.get_connection_stats()
-        stream_stats = price_stream_manager.get_stats()
-
-        print("✅ Integration stats:")
-        print(f"   - WebSocket connections: {ws_stats['total_connections']}")
-        print(f"   - Stream manager running: {stream_stats['manager_running']}")
-
-        return True
-
-    except Exception as e:
-        print(f"❌ WebSocket integration test failed: {e}")
-        return False
-
-
-async def main():
-    """メイン実行関数"""
-    print("🚀 Starting comprehensive price streaming system tests...\n")
-
-    # テスト実行
-    test_results = []
-
-    tests = [
-        ("Price Streamer Import", test_price_streamer_import),
-        ("Price Manager Lifecycle", test_price_manager_lifecycle),
-        ("Message Broadcasting", test_message_broadcasting),
-        ("Streaming Routes Import", test_streaming_routes_import),
-        ("Main App Integration", test_main_app_streaming_integration),
-        ("Binance Streamer Config", test_binance_streamer_config),
-        ("WebSocket Integration", test_websocket_integration),
-    ]
-
-    for test_name, test_func in tests:
-        try:
-            result = await test_func()
-            test_results.append((test_name, result))
-        except Exception as e:
-            print(f"❌ Test '{test_name}' encountered an error: {e}")
-            test_results.append((test_name, False))
-
-    # 結果表示
-    print("\n📊 Comprehensive Test Results:")
-    passed = 0
-    for test_name, result in test_results:
-        status = "✅ PASS" if result else "❌ FAIL"
-        print(f"   - {test_name}: {status}")
-        if result:
-            passed += 1
-
-    success_rate = passed / len(test_results) * 100
-    print(f"\n🎯 Success Rate: {passed}/{len(test_results)} ({success_rate:.1f}%)")
-
-    if success_rate >= 80:
-        print("\n🎉 Price streaming system is working excellently!")
-        print("   Ready for real-time cryptocurrency price distribution.")
-        print("   ✅ Binance WebSocket integration")
-        print("   ✅ Multi-channel broadcasting")
-        print("   ✅ Automatic price caching")
-        print("   ✅ Admin control APIs")
-    elif success_rate >= 60:
-        print("\n✅ Price streaming system is mostly working.")
-        print("   Minor issues detected but core functionality is intact.")
-    else:
-        print("\n⚠️  Some critical issues detected.")
-        print("   Please review the failing tests before proceeding.")
-
-    return success_rate >= 80
-
-
-if __name__ == "__main__":
-    success = asyncio.run(main())
-    sys.exit(0 if success else 1)
